EnemyBase wip

+++ Assets/Labor-Man/Scripts/Enemy/EnemyBase.cs
@@ -8,7 +8,7 @@ public class EnemyBase : MonoBehaviour,IDamageble
 {
     public float Speed => _enemySpeed;
     Transform _myTransform = default;
-    public float EnemyHp => _enemyHp; 
+    public int EnemyHp => _enemyHp; 
     SpriteRenderer _sp;
 
     [SerializeField]
